StatusCode as int32

+++ response.go
@@ -9,17 +9,17 @@ type TxResponse struct {
 	porthos.Response
 }
 
-func (r *TxResponse) JSON(statusCode int16, body interface{}) {
+func (r *TxResponse) JSON(statusCode int32, body interface{}) {
 	defer r.StartSegment("Response time").End()
 	r.Response.JSON(statusCode, body)
 }
 
-func (r *TxResponse) Raw(statusCode int16, contentType string, body []byte) {
+func (r *TxResponse) Raw(statusCode int32, contentType string, body []byte) {
 	defer r.StartSegment("Response time").End()
 	r.Response.Raw(statusCode, contentType, body)
 }
 
-func (r *TxResponse) Empty(statusCode int16) {
+func (r *TxResponse) Empty(statusCode int32) {
 	defer r.StartSegment("Response time").End()
 	r.Response.Empty(statusCode)
 }
